setup app to work on a sub path
