+++ .env
@@ -1,0 +1,1 @@
+NEXT_PUBLIC_BASE_PATH=/usdc-bridge
