+++ package.json
@@ -5,6 +5,7 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
+    "build:static": "next build && next export -o build",
     "start": "next start",
     "lint": "next lint",
     "deployDev": "NEXT_PUBLIC_NETWORK=testnet DEPLOY=deploy npm run build && touch out/.nojekyll && cp CNAME_DEV out/ && mv out/CNAME_DEV out/CNAME && npx push-dir --dir=out --branch=gh-pages --cleanup --verbose",
