+++ src/pages/app.module.scss
@@ -81,7 +81,7 @@
           text-decoration: underline;
         }
 
-        a[href="https://stable.portalbridge.com/"]
+        a[href="https://www.portalbridge.com/usdc-bridge"]
         {
           text-decoration: underline;
         }
